refactor (parser.js): relocate to package root
* update tests and paths
* relocate test spec to match pathing
